--- src/main/java/com/project/cheerha/common/exeption/ErrorCode.java
@@ -1,17 +1,0 @@
-package com.project.cheerha.common.exeption;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@RequiredArgsConstructor
-public enum ErrorCode {
-
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "존재하지 않는 사용자입니다."),
-    WRONG_EMAIL_OR_PASSWORD(HttpStatus.NOT_FOUND, "이메일이나 패스워드가 잘못되었습니다."),
-    ALREADY_EXIST_EMAIL(HttpStatus.BAD_REQUEST, "이미 존재하는 이메일입니다.");
-
-    private final HttpStatus status;
-    private final String message;
-}
