+++ src/main/java/com/project/cheerha/domain/auth/controller/AuthController.java
@@ -12,6 +12,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -38,8 +39,9 @@ public class AuthController {
 
     @PostMapping("/logout")
     public ResponseEntity<CreateLogoutResponseDto> logout(
-
+        @RequestHeader("Authorization") String authHeader
     ){
-        return ResponseEntity.status(HttpStatus.OK).body(authService.logout());
+        return ResponseEntity.status(HttpStatus.OK).body(authService.logout(authHeader));
     }
+
 }
